Moves accordion-item-body to wrapper file structure
